requirements.txt
fitz==0.0.1.dev2
numpy==1.22.3
opencv_python==4.11.0.86
pdf2image==1.17.0
pdfplumber==0.11.5
Pillow==9.1.0
Pillow==11.1.0
pytesseract==0.3.13
python_docx==1.1.2
python_magic==0.4.27
spacy==3.8.4
